Split shell structured delete tests
